+++ CHANGES.rst
@@ -4,6 +4,13 @@ Change Log
 prawcore follows `semantic versioning <http://semver.org/>`_ with the exception
 that deprecations will not be announced by a minor release.
 
+Unreleased
+----------
+
+**Added**
+
+* Added ``data`` parameter to ``Session.request``.
+
 0.0.6 (2016-03-14)
 
 **Fixed**